Separation of conserns
I separated the interop API calls and their helper methods. The helper
method are now in the MonoLibUsbApiHelpers.cs file.
I also reordered the API method definitions to reflect the original C++
header. It's better for the future updates this way.
